document control organization
* simplified control routing
* better use of context
