+++ apps/extension/src/control/internal-message.ts
@@ -1,0 +1,20 @@
+export interface InternalMessage<Type extends string, Req, Res> {
+  type: Type;
+  request: Req;
+  response: Res;
+}
+
+export interface InternalRequest<M extends InternalMessage<string, unknown, unknown>> {
+  type: M['type'];
+  request: M['request'];
+}
+
+export type InternalResponse<M extends InternalMessage<string, unknown, unknown>> =
+  | {
+      type: M['type'];
+      data: M['response'];
+    }
+  | {
+      type: M['type'];
+      error: unknown;
+    };
