+++ apps/extension/src/ctx/index.ts
@@ -1,3 +1,3 @@
-export * from './tx-approval';
-export * from './wallet-id';
+export * from './popup';
+export * from './local-storage';
 export * from './offscreen';
