+++ apps/extension/src/ctx/popup.ts
@@ -1,0 +1,4 @@
+import type { GetTxApprovalFn } from '@penumbra-zone/router';
+import { popupControl } from '../control/popup';
+
+export const getTxApproval: GetTxApprovalFn = popupControl.txApproval;
