+++ apps/extension/src/hooks/onboarding.ts
@@ -3,6 +3,7 @@ import { passwordSelector } from '../state/password';
 import { generateSelector } from '../state/seed-phrase/generate';
 import { importSelector } from '../state/seed-phrase/import';
 import { walletsSelector } from '../state/wallets';
+import { blockCacheControl } from '../control/services';
 
 // Saves hashed password, uses that hash to encrypt the seed phrase
 // and then saves that to session + local storage
@@ -18,6 +19,6 @@ export const useOnboardingSave = () => {
     await setPassword(plaintextPassword);
 
     await addWallet({ label: 'Wallet #1', seedPhrase });
-    void chrome.runtime.sendMessage('PENUMBRA_SYNC_BLOCKS');
+    void blockCacheControl.syncBlocks();
   };
 };
