+++ apps/extension/src/popup.tsx
@@ -5,11 +5,12 @@ import { RouterProvider } from 'react-router-dom';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
 import '@penumbra-zone/ui/styles/globals.css';
-import { popupMsgHandler } from './routes/popup/internal-msg-handler';
+
+import { popupControlHandler } from './control/popup';
 
 export const queryClient = new QueryClient();
 
-chrome.runtime.onMessage.addListener(popupMsgHandler);
+chrome.runtime.onMessage.addListener(popupControlHandler);
 
 const startPopup = () => {
   const rootElement = document.getElementById('popup-root') as HTMLDivElement;
