+++ apps/extension/src/state/tx-approval.ts
@@ -1,12 +1,11 @@
 import { AllSlices, SliceCreator } from './index';
-import { MessageResponder } from '@penumbra-zone/types/src/internal-msg/shared';
-import { TxApproval } from '@penumbra-zone/types/src/internal-msg/tx-approval';
+import { TxApprovalResponder } from '../control/popup';
 
 export interface TxApprovalSlice {
   // Is AuthorizeRequest-jsonified. Zustand+Immer does not allow for storing classes and Typescript does not allow for JsValue.
   tx: object | undefined;
   // Holding the message responder function. Service worker will be "awaiting" the call of this.
-  responder: MessageResponder<TxApproval> | undefined;
+  responder: TxApprovalResponder | undefined;
 }
 
 export const createTxApprovalSlice: SliceCreator<TxApprovalSlice> = () => ({
