+++ packages/router/src/grpc/view-protocol-server/witness-and-build.ts
@@ -1,5 +1,5 @@
 import type { Impl } from '.';
-import { servicesCtx, offscreenCtx } from '../../ctx';
+import { servicesCtx, buildActionCtx } from '../../ctx';
 
 import { buildParallel, witness } from '@penumbra-zone/wasm-ts';
 
@@ -7,7 +7,7 @@ import { ConnectError, Code } from '@connectrpc/connect';
 
 export const witnessAndBuild: Impl['witnessAndBuild'] = async (req, ctx) => {
   const services = ctx.values.get(servicesCtx);
-  const offscreenClient = ctx.values.get(offscreenCtx);
+  const buildAction = ctx.values.get(buildActionCtx);
   if (!req.authorizationData)
     throw new ConnectError('No authorization data in request', Code.InvalidArgument);
   if (!req.transactionPlan) throw new ConnectError('No tx plan in request', Code.InvalidArgument);
@@ -20,7 +20,7 @@ export const witnessAndBuild: Impl['witnessAndBuild'] = async (req, ctx) => {
 
   const witnessData = witness(req.transactionPlan, sct);
 
-  const batchActions = await offscreenClient.buildAction(req, witnessData, fullViewingKey);
+  const batchActions = await buildAction(req, witnessData, fullViewingKey);
 
   const transaction = buildParallel(
     batchActions,
