--- packages/types/src/internal-msg/tx-approval.ts
@@ -1,5 +1,0 @@
-import { InternalMessage } from './shared';
-import { JsonValue } from '@bufbuild/protobuf';
-
-// JsonValue === AuthorizeRequest
-export type TxApproval = InternalMessage<'TX-APPROVAL', JsonValue, boolean>;
